split stylesheets
