+++ src/components/organisms/ModalForm/components/FieldSet/styles.ts
@@ -1,0 +1,8 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  contentContainer: {
+    paddingTop: 8,
+    gap: 8,
+  },
+});
